# apps/docs/api-reference/openapi.yaml
openapi: 3.1.0
info:
  title: Authlane Control Plane API
  version: 1.0.0
  description: Manages SaaS tenant policy, end-user connections, credentials, and tool definitions. Provider tool traffic does not pass through this API.
servers:
  - url: https://authlane.example.com
  - url: http://localhost:3000
security:
  - apiKey: []
paths:
...
  /api/v1/users/{externalUserId}/connections:
    get:
      operationId: listConnections
      summary: List effective connection statuses
      x-authlane-scope: connections:read
      parameters:
        - $ref: '#/components/parameters/ExternalUserId'
      responses:
        '200':
          description: Connected and disconnected service statuses
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionsResult'
  /api/v1/users/{externalUserId}/capabilities:
    get:
      operationId: getCapabilities
      summary: Get statuses and tool definitions in one hot-read snapshot
      x-authlane-scope: connections:read
      parameters:
        - $ref: '#/components/parameters/ExternalUserId'
        - $ref: '#/components/parameters/ToolFormat'
      responses:
        '200':
          description: Capability snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CapabilitiesResult'
  /api/v1/users/{externalUserId}/tools:
    get:
      operationId: listTools
      summary: List definitions for currently usable connections
      x-authlane-scope: connections:read
      parameters:
        - $ref: '#/components/parameters/ExternalUserId'
        - $ref: '#/components/parameters/ToolFormat'
      responses:
        '200':
          description: MCP tools or OpenAI functions with version hash
  /api/v1/users/{externalUserId}/connections/{serviceId}/credentials:
    get:
      operationId: getCredentials
      summary: Get access-only credential material
      description: Server-only, audited, no-store response. OAuth refresh tokens are never returned.
      x-authlane-scope: credentials:read
      parameters:
        - $ref: '#/components/parameters/ExternalUserId'
        - $ref: '#/components/parameters/ServiceId'
      responses:
        '200':
          description: Credential material
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialResult'
        '409':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      bearerFormat: Authlane scoped API key
  parameters:
    ExternalUserId:
      name: externalUserId
      in: path
      required: true
      schema:
        type: string
        maxLength: 255
    ServiceId:
      name: serviceId
      in: path
      required: true
      schema:
        type: string
        pattern: '^[a-z0-9-]+$'
    ToolFormat:
      name: format
      in: query
      schema:
        type: string
        enum: [mcp, openai]
        default: mcp
  schemas:
    Error:
      type: object
      required: [message, code]
      properties:
        message:
          type: string
        code:
          type: string
        hint:
          type: string
        docUrl:
          type: string
          format: uri
    Connection:
      type: object
      required: [serviceId, status, connected, expiresAt]
      properties:
        serviceId:
          type: string
        status:
          type: string
          enum: [disconnected, pending, connected, expired, error]
        connected:
          type: boolean
        expiresAt:
          type: [string, 'null']
          format: date-time
        connectedAt:
          type: [string, 'null']
          format: date-time
        lastCheckedAt:
          type: [string, 'null']
          format: date-time
        errorCode:
          type: [string, 'null']
    ConnectionsResult:
      type: object
      required: [data, error]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Connection'
        error:
          type: 'null'
    CapabilitiesResult:
      type: object
      required: [data, error]
      properties:
        data:
          type: object
          required: [externalUserId, format, version, services]
          properties:
            externalUserId:
              type: string
            format:
              type: string
              enum: [mcp, openai]
            version:
              type: string
            services:
              type: array
              items:
                allOf:
                  - $ref: '#/components/schemas/Connection'
                  - type: object
                    required: [tools]
                    properties:
                      tools:
                        type: array
                        items:
                          type: object
        error:
          type: 'null'
    CredentialResult:
      type: object
      required: [data, error]
      properties:
        data:
          oneOf:
            - type: object
              required: [type, accessToken, tokenType, scopes, expiresAt]
              properties:
                type:
                  const: oauth2
                accessToken:
                  type: string
                tokenType:
                  type: string
                scopes:
                  type: array
                  items:
                    type: string
                expiresAt:
                  type: [string, 'null']
            - type: object
              required: [type, apiKey]
              properties:
                type:
                  const: api_key
                apiKey:
                  type: string
                apiSecret:
                  type: string
            - type: object
              required: [type, headers]
              properties:
                type:
                  const: header
                headers:
                  type: object
                  additionalProperties:
                    type: string
        error:
          type: 'null'
...
  responses:
...
    Error:
      description: Authlane error
      content:
        application/json:
          schema:
            type: object
            required: [data, error]
            properties:
              data:
                type: 'null'
              error:
                $ref: '#/components/schemas/Error'